stop jekyll serve from constant rebuilding

+++ jekyll/_plugins/update_yml.rb
@@ -10,7 +10,8 @@ module Jekyll
       json_data = JSON.parse(File.read(data_file))
       updated_yaml = json_data.to_yaml
       target_file = File.join(site.source, '_data', 'data.yml')
-      File.write(target_file, updated_yaml)
+      existing = File.exist?(target_file) ? File.read(target_file) : nil
+      File.write(target_file, updated_yaml) unless existing == updated_yaml
       # puts "\e[34mupdate_yml.rb plugin took #{Time.now - start_time} seconds.\e[0m"
     end
